Apply ruff formatting fixes and fix test file gitignore pattern
Bumps the pre-commit ruff rev to v0.14.4 and reformats/simplifies code across
sim/spectra/viz/runner modules to match. Also narrows the .gitignore test*
pattern to /test_* so it only excludes root-level scratch files instead of
accidentally excluding tests/test_*.py, surfacing several previously-hidden
test files.

+++ nvision/metrics/series.py
@@ -45,7 +45,7 @@ def _downsample_indices(n: int, max_points: int) -> list[int]:
     ratio = (n - 1) / head
     for k in range(remaining):
         t = (k + 1) / (remaining + 1)
-        indices.add(int(round(head * ratio**t)))
+        indices.add(round(head * ratio**t))
     indices.add(n - 1)
     return sorted(i for i in indices if 0 <= i < n)
